refactor: split pkg/flare to remove unneeded deps
