+++ pkg/flare/archive.go
@@ -11,17 +11,15 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"expvar"
 	"fmt"
 	"io"
 	"net/http"
 	"path/filepath"
-	"regexp"
 	"sort"
-	"strings"
 	"time"
 
 	"github.com/fatih/color"
+	"gopkg.in/yaml.v2"
 
 	sysprobeclient "github.com/DataDog/datadog-agent/cmd/system-probe/api/client"
 	flaretypes "github.com/DataDog/datadog-agent/comp/core/flare/types"
@@ -31,21 +29,15 @@ import (
 	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
 	"github.com/DataDog/datadog-agent/pkg/diagnose"
 	"github.com/DataDog/datadog-agent/pkg/diagnose/diagnosis"
+	"github.com/DataDog/datadog-agent/pkg/flare/common"
+	"github.com/DataDog/datadog-agent/pkg/flare/priviledged"
 	"github.com/DataDog/datadog-agent/pkg/status/health"
 	systemprobeStatus "github.com/DataDog/datadog-agent/pkg/status/systemprobe"
 	"github.com/DataDog/datadog-agent/pkg/util/cloudproviders"
 	"github.com/DataDog/datadog-agent/pkg/util/ecs"
 	"github.com/DataDog/datadog-agent/pkg/util/installinfo"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
-
-	"gopkg.in/yaml.v2"
 )
-
-// Match .yaml and .yml to ship configuration files in the flare.
-var cnfFileExtRx = regexp.MustCompile(`(?i)\.ya?ml`)
-
-// searchPaths is a list of path where to look for checks configurations
-type searchPaths map[string]string
 
 // getProcessAPIAddress is an Alias to GetProcessAPIAddressPort using Datadog config
 func getProcessAPIAddressPort() (string, error) {
@@ -70,7 +62,7 @@ func ExtraFlareProviders(diagnoseDeps diagnose.SuitesDeps) []*flaretypes.FlareFi
 		flaretypes.NewFiller(getRegistryJSON),
 		flaretypes.NewFiller(getVersionHistory),
 		flaretypes.NewFiller(getWindowsData),
-		flaretypes.NewFiller(GetExpVar),
+		flaretypes.NewFiller(common.GetExpVar),
 		flaretypes.NewFiller(provideInstallInfo),
 		flaretypes.NewFiller(provideAuthTokenPerm),
 		flaretypes.NewFiller(provideDiagnoses(diagnoseDeps)),
@@ -82,7 +74,7 @@ func ExtraFlareProviders(diagnoseDeps diagnose.SuitesDeps) []*flaretypes.FlareFi
 	telemetryURL := fmt.Sprintf("http://127.0.0.1:%s/telemetry", pkgconfigsetup.Datadog().GetString("expvar_port"))
 
 	for filename, fromFunc := range map[string]func() ([]byte, error){
-		"envvars.log":         GetEnvVars,
+		"envvars.log":         common.GetEnvVars,
 		"health.yaml":         getHealth,
 		"go-routine-dump.log": func() ([]byte, error) { return getHTTPCallContent(pprofURL) },
 		"telemetry.log":       func() ([]byte, error) { return getHTTPCallContent(telemetryURL) },
@@ -199,75 +191,10 @@ func getRegistryJSON(fb flaretypes.FlareBuilder) error {
 	return nil
 }
 
-// GetLogFiles copies log files to the flare archive.
-func GetLogFiles(fb flaretypes.FlareBuilder, logFileDir string) {
-	log.Flush()
-
-	fb.CopyDirToWithoutScrubbing(filepath.Dir(logFileDir), "logs", func(path string) bool { //nolint:errcheck
-		if filepath.Ext(path) == ".log" || getFirstSuffix(path) == ".log" {
-			return true
-		}
-		return false
-	})
-}
-
-// GetExpVar copies expvar files to the flare archive.
-func GetExpVar(fb flaretypes.FlareBuilder) error {
-	variables := make(map[string]interface{})
-	expvar.Do(func(kv expvar.KeyValue) {
-		variable := make(map[string]interface{})
-		json.Unmarshal([]byte(kv.Value.String()), &variable) //nolint:errcheck
-		variables[kv.Key] = variable
-	})
-
-	// The callback above cannot return an error.
-	// In order to properly ensure error checking,
-	// it needs to be done in its own loop
-	for key, value := range variables {
-		yamlValue, err := yaml.Marshal(value)
-		if err != nil {
-			return err
-		}
-
-		err = fb.AddFile(filepath.Join("expvar", key), yamlValue)
-		if err != nil {
-			return err
-		}
-	}
-
-	apmDebugPort := pkgconfigsetup.Datadog().GetInt("apm_config.debug.port")
-	f := filepath.Join("expvar", "trace-agent")
-	resp, err := http.Get(fmt.Sprintf("https://127.0.0.1:%d/debug/vars", apmDebugPort))
-	if err != nil {
-		return fb.AddFile(f, []byte(fmt.Sprintf("Error retrieving vars: %v", err)))
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		slurp, err := io.ReadAll(resp.Body)
-		if err != nil {
-			return err
-		}
-		return fb.AddFile(f, []byte(fmt.Sprintf("Got response %s from /debug/vars:\n%s", resp.Status, slurp)))
-	}
-	var all map[string]interface{}
-	if err := json.NewDecoder(resp.Body).Decode(&all); err != nil {
-		return fmt.Errorf("error decoding trace-agent /debug/vars response: %v", err)
-	}
-	v, err := yaml.Marshal(all)
-	if err != nil {
-		return err
-	}
-	return fb.AddFile(f, v)
-}
-
-func getSystemProbeSocketPath() string {
-	return pkgconfigsetup.SystemProbe().GetString("system_probe_config.sysprobe_socket")
-}
-
 func getSystemProbeStats() ([]byte, error) {
 	// TODO: (components) - Temporary until we can use the status component to extract the system probe status from it.
 	stats := map[string]interface{}{}
-	systemprobeStatus.GetStatus(stats, getSystemProbeSocketPath())
+	systemprobeStatus.GetStatus(stats, priviledged.GetSystemProbeSocketPath())
 	sysProbeBuf, err := yaml.Marshal(stats["systemProbeStats"])
 	if err != nil {
 		return nil, err
@@ -277,13 +204,13 @@ func getSystemProbeStats() ([]byte, error) {
 }
 
 func getSystemProbeTelemetry() ([]byte, error) {
-	sysProbeClient := sysprobeclient.Get(getSystemProbeSocketPath())
+	sysProbeClient := sysprobeclient.Get(priviledged.GetSystemProbeSocketPath())
 	url := sysprobeclient.URL("/telemetry")
 	return getHTTPData(sysProbeClient, url)
 }
 
 func getSystemProbeConfig() ([]byte, error) {
-	sysProbeClient := sysprobeclient.Get(getSystemProbeSocketPath())
+	sysProbeClient := sysprobeclient.Get(priviledged.GetSystemProbeSocketPath())
 	url := sysprobeclient.URL("/config")
 	return getHTTPData(sysProbeClient, url)
 }
@@ -302,40 +229,6 @@ func getProcessAgentFullConfig() ([]byte, error) {
 		return []byte("error: process-agent is not running or is unreachable\n"), nil
 	}
 	return bytes, nil
-}
-
-// GetConfigFiles copies configuration files to the flare archive.
-func GetConfigFiles(fb flaretypes.FlareBuilder, confSearchPaths map[string]string) {
-	for prefix, filePath := range confSearchPaths {
-		fb.CopyDirTo(filePath, filepath.Join("etc", "confd", prefix), func(path string) bool { //nolint:errcheck
-			// ignore .example file
-			if filepath.Ext(path) == ".example" {
-				return false
-			}
-
-			firstSuffix := []byte(getFirstSuffix(path))
-			ext := []byte(filepath.Ext(path))
-			if cnfFileExtRx.Match(firstSuffix) || cnfFileExtRx.Match(ext) {
-				return true
-			}
-			return false
-		})
-	}
-
-	if pkgconfigsetup.Datadog().ConfigFileUsed() != "" {
-		mainConfpath := pkgconfigsetup.Datadog().ConfigFileUsed()
-		confDir := filepath.Dir(mainConfpath)
-
-		// zip up the config file that was actually used, if one exists
-		fb.CopyFileTo(mainConfpath, filepath.Join("etc", "datadog.yaml")) //nolint:errcheck
-
-		// figure out system-probe file path based on main config path, and use best effort to include
-		// system-probe.yaml to the flare
-		fb.CopyFileTo(filepath.Join(confDir, "system-probe.yaml"), filepath.Join("etc", "system-probe.yaml")) //nolint:errcheck
-
-		// use best effort to include security-agent.yaml to the flare
-		fb.CopyFileTo(filepath.Join(confDir, "security-agent.yaml"), filepath.Join("etc", "security-agent.yaml")) //nolint:errcheck
-	}
 }
 
 func getChecksFromProcessAgent(fb flaretypes.FlareBuilder, getAddressPort func() (url string, err error)) {
@@ -537,10 +430,6 @@ func getHTTPCallContent(url string) ([]byte, error) {
 	}
 
 	return data, nil
-}
-
-func getFirstSuffix(s string) string {
-	return filepath.Ext(strings.TrimSuffix(s, filepath.Ext(s)))
 }
 
 // functionOutputToBytes runs a given function and returns its output in a byte array
