+++ pkg/flare/archive_test.go
@@ -42,42 +42,6 @@ func TestGoRoutines(t *testing.T) {
 	assert.Equal(t, expected, string(content))
 }
 
-func TestIncludeSystemProbeConfig(t *testing.T) {
-	configmock.NewFromFile(t, "./test/datadog-agent.yaml")
-	// create system-probe.yaml file because it's in .gitignore
-	_, err := os.Create("./test/system-probe.yaml")
-	require.NoError(t, err, "couldn't create system-probe.yaml")
-	defer os.Remove("./test/system-probe.yaml")
-
-	mock := flarehelpers.NewFlareBuilderMock(t, false)
-	GetConfigFiles(mock, searchPaths{"": "./test/confd"})
-
-	mock.AssertFileExists("etc", "datadog.yaml")
-	mock.AssertFileExists("etc", "system-probe.yaml")
-}
-
-func TestIncludeConfigFiles(t *testing.T) {
-	configmock.New(t)
-
-	mock := flarehelpers.NewFlareBuilderMock(t, false)
-	GetConfigFiles(mock, searchPaths{"": "./test/confd"})
-
-	mock.AssertFileExists("etc/confd/test.yaml")
-	mock.AssertFileExists("etc/confd/test.Yml")
-	mock.AssertNoFileExists("etc/confd/not_included.conf")
-}
-
-func TestIncludeConfigFilesWithPrefix(t *testing.T) {
-	configmock.New(t)
-
-	mock := flarehelpers.NewFlareBuilderMock(t, false)
-	GetConfigFiles(mock, searchPaths{"prefix": "./test/confd"})
-
-	mock.AssertFileExists("etc/confd/prefix/test.yaml")
-	mock.AssertFileExists("etc/confd/prefix/test.Yml")
-	mock.AssertNoFileExists("etc/confd/prefix/not_included.conf")
-}
-
 func createTestFile(t *testing.T, filename string) string {
 	path := filepath.Join(t.TempDir(), filename)
 	require.NoError(t, os.WriteFile(path, []byte("mockfilecontent"), os.ModePerm))
