+++ pkg/flare/securityagent/test/compliance.d/cis-docker.yaml
